championkids-api/requirements.txt
# ChampionKids API — production dependencies
# Pin exact versions to ensure reproducible builds.

fastapi==0.111.1
uvicorn[standard]==0.30.1
motor==3.4.0
pymongo==4.7.3
pydantic[email]==2.7.4
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
APScheduler==3.10.4
python-dateutil==2.9.0
pytz==2024.1
slowapi==0.1.9
sentry-sdk[fastapi]==2.6.0
stripe==9.12.0
httpx==0.27.0
google-auth==2.30.0
PyJWT==2.8.0
